requirements.txt
langchain 
numpy 
streamlit 
langchain-community 
streamlit 
PyPDF2 
unstructured 
nltk 
sentence-transformers 
langchain-huggingface 
faiss-cpu 
python-docx
fastapi
uvicorn
jinja2